Fix a race condition when evaluating on the root context
The root runner's Evaluator is shared across goroutines,
so side effects on the CallStack are not goroutine safe.

We solve this problem by moving the CallStack into a Scope
for each evaluation.
